// src/modules/area-estrategia-desarrollo-comercial/analisis_venues/application/dto/create-analisis-venue.dto.ts
export class CreateAnalisisVenueDto {
  nombre?: string;
  departamento?: string;
  distrito?: string;
  direccion?: string;
  celular?: string;
  correo?: string;
  capacidad_personas?: number | string;
  capacidadPersonas?: number | string;
  estado?: string;
  sitio_web?: string;
  sitioWeb?: string;
  detalles?: string;
}